+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/StreamDimens.kt
@@ -13,6 +13,13 @@ import androidx.compose.ui.unit.dp
  * @param channelInfoUserItemWidth The width of a member tile in the channel info dialog.
  * @param channelInfoUserItemHorizontalPadding The padding inside a member tile in the channel info dialog.
  * @param channelInfoUserItemAvatarSize The size of a member avatar in the channel info dialog.
+ * @param attachmentsContentImageWidth The width of
+ * @param attachmentsContentImageHeight The height of
+ * @param attachmentsContentGiphyWidth The with of
+ * @param attachmentsContentGiphyHeight The height of
+ * @param attachmentsContentLinkWidth The with of link
+ * @param attachmentsContentFileWidth The width of file
+ * @param attachmentsContentFileUploadWidth The Width of
  */
 @Immutable
 public data class StreamDimens(
@@ -22,6 +29,13 @@ public data class StreamDimens(
     public val channelInfoUserItemWidth: Dp,
     public val channelInfoUserItemHorizontalPadding: Dp,
     public val channelInfoUserItemAvatarSize: Dp,
+    public val attachmentsContentImageWidth: Dp,
+    public val attachmentsContentImageHeight: Dp,
+    public val attachmentsContentGiphyWidth: Dp,
+    public val attachmentsContentGiphyHeight: Dp,
+    public val attachmentsContentLinkWidth: Dp,
+    public val attachmentsContentFileWidth: Dp,
+    public val attachmentsContentFileUploadWidth: Dp,
 ) {
     public companion object {
         public fun defaultDimens(): StreamDimens = StreamDimens(
@@ -31,6 +45,13 @@ public data class StreamDimens(
             channelInfoUserItemWidth = 80.dp,
             channelInfoUserItemHorizontalPadding = 8.dp,
             channelInfoUserItemAvatarSize = 64.dp,
+            attachmentsContentImageWidth = 250.dp,
+            attachmentsContentImageHeight = 200.dp,
+            attachmentsContentGiphyWidth = 250.dp,
+            attachmentsContentGiphyHeight = 200.dp,
+            attachmentsContentLinkWidth = 250.dp,
+            attachmentsContentFileWidth = 250.dp,
+            attachmentsContentFileUploadWidth = 250.dp,
         )
     }
 }